pretend.rip/tools/requirements.txt
discord.py
jishaku
psutil
xxhash
pillow
redis
munch
tornado
beautifulsoup4
humanize
pomice
shazamio
typing_extensions
asyncpg
yt-dlp
pyppeteer
colorgram.py
python-dateutil
pydantic
humanfriendly
cashews[redis]
python-magic
ffmpeg-python
discord_ios
sublist3r
whapi
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus
discord-paginator @ git+https://github.com/prettylittlelies/discord_paginator
bcrypt
arrow
aiohttp
aiofiles
colorama
cycler
httptools
oauthlib
python-dotenv
pytz
Quart
shazamio
tuuid
uvloop
uwuify
google.generativeai
